Change exception thrown by getKey in ISOComponent
Throwing ISOException checked exception in ISOComponent forces every usage
of the method to catch it. In many use cases developers know that this
exception cannot occur but still need to make empty catch block.
Additionally the usual type of exception in composite pattern in java is
IllegalStateException.

This commit begins the series of transforming ISOComponent methods to
throw IllegalStateException (or IllegalArgumentException) instead of
ISOException.